Начало работы с Domain слоем. Изучения основ чистой архетектуры. Создание UseCase`ов и моделей для приложения. Тем самым убирая всю логику из Activity (Presentation слоя). Тем самым повысив читаемость кода.
