+++ app/src/main/java/com/example/cleanarchitectureapp/domain/models/SaveUserNameParam.kt
@@ -1,0 +1,3 @@
+package com.example.cleanarchitectureapp.domain.models
+
+data class SaveUserNameParam (val firstName: String)
