+++ app/src/main/java/com/example/cleanarchitectureapp/domain/models/UserName.kt
@@ -1,0 +1,6 @@
+package com.example.cleanarchitectureapp.domain.models
+//Класс содержаший имя пользователя
+data class UserName(
+    val firstName: String,
+    val lastName: String
+)
