+++ app/src/main/java/com/example/cleanarchitectureapp/domain/usecase/GetUserNameUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.cleanarchitectureapp.domain.usecase
+
+import com.example.cleanarchitectureapp.domain.models.UserName
+
+class GetUserNameUseCase {
+    operator fun invoke(): UserName = UserName(firstName = "Leon", lastName = "Uraltsev")
+}
